+++ client/package.json
@@ -33,6 +33,7 @@
     ]
   },
   "devDependencies": {
-    "@material-ui/core": "^4.11.0"
+    "@material-ui/core": "^4.11.0",
+    "react-router-dom": "^5.2.0"
   }
 }
