+++ client/src/component/mainpage/Banner.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import tempImg from 'banner/1.gif';
+const Banner = () => {
+  return (
+    <div>
+      <section className="main-banner">
+        <img src={tempImg}></img>
+      </section>
+    </div>
+  );
+};
+
+export default Banner;
