+++ client/src/component/share/Product.scss
@@ -4,6 +4,7 @@
     justify-content: space-between;
 
   &-header {
+    font-family:'BMDOHYEON';
     display: flex;
     align-items: center;
     justify-content: space-between;
@@ -18,6 +19,7 @@
 }
 
 .product-item {
+    font-size:0.8em;
     width:45%;
   display: flex;
   flex-direction: column;
